Improve organization and simplify compilation
