+++ .babelrc
@@ -1,0 +1,7 @@
+{
+  "presets": ["react", "stage-0"],
+  "plugins": [
+    ["transform-react-jsx", { "pragma": "h" }],
+    "transform-runtime"
+  ]
+}
