+++ package.json
@@ -5,11 +5,14 @@
   "module": "index.js",
   "main": "dist/index.js",
   "scripts": {
-    "build": "node ./compileComponents.js && webpack",
+    "build": "npm run build:umd && npm run build:es",
+    "build:umd": "cross-env NODE_ENV=production webpack --config webpack.config.umd.js",
+    "build:es": "cross-env NODE_ENV=production babel . --out-dir . --extensions \".jsx\" --ignore node_modules,sample && npm run copycss",
+    "copycss": "node copyCss.js",
+    "precommit": "lint-staged",
     "prepublish": "npm run build && npm run test ",
     "test": "npm run build && npm run lint",
-    "lint": "eslint src/",
-    "lint-staged": "lint-staged"
+    "lint": "eslint ."
   },
   "repository": {
     "type": "git",
@@ -47,7 +50,6 @@
     "preact": "*"
   },
   "devDependencies": {
-    "babel": "^6.5.2",
     "babel-cli": "^6.24.0",
     "babel-core": "^6.21.0",
     "babel-eslint": "^7.2.2",
@@ -70,25 +72,19 @@
     "babili-webpack-plugin": "0.1.2",
     "clean-webpack-plugin": "^0.1.15",
     "cp": "^0.2.0",
+    "cross-env": "^5.0.2",
     "eslint": "^4.0.0",
     "eslint-plugin-react": "^7.0.0",
+    "husky": "^0.14.3",
     "lint-staged": "^4.0.0",
     "mkdirp": "^0.5.1",
-    "pre-commit": "^1.2.2",
     "preact": "^8.2.1",
     "prettier": "^1.3.1",
+    "rimraf": "^2.6.1",
     "webpack": "^3.0.0",
     "webpack-shell-plugin": "^0.5.0"
   },
   "lint-staged": {
-    "index.js": [
-      "prettier --write",
-      "git add"
-    ],
-    "**/*.jsx": [
-      "prettier --write",
-      "git add"
-    ]
-  },
-  "pre-commit": "lint-staged"
+    "**/*.{js,jsx,json,css}": ["prettier --write", "git add"]
+  }
 }
